feat(chat): 새  채팅페이지로 router  갈아끼우고 컴포넌트 분리
