+++ src/pages/chat/components/room/index.ts
@@ -1,4 +1,2 @@
-export { ChatRoomInsideSection } from './ChatRoomInsideSection';
-export { ChatRoomSection } from './ChatRoomSection';
+export { ChatRoomInside } from './ChatRoomInside';
 export { ChatRoomList } from './ChatRoomList';
-export { ChatRoomListItem } from './ChatRoomListItem';
