+++ src/pages/chat/components/user/index.ts
@@ -1,3 +1,1 @@
-export { ChatUserSection } from './ChatUserSection';
 export { ChatUserList } from './ChatUserList';
-export { ChatUserListItem } from './ChatUserListItem';
